# docs/swagger.yml
openapi: 3.0.3
info:
  title: ACM UIUC Membership API
  version: "1.0.0"
  contact:
    name: ACM Infrastructure Team
    email: infra@acm.illinois.edu

paths:
  /api/v1/healthz:
    get:
      summary: Ping the API
      operationId: ping

      responses:
        200:
          description: OK


      x-amazon-apigateway-auth:
        type: NONE

      x-amazon-apigateway-integration:
        responses:
          default:
            statusCode: 200
        passthroughBehavior: when_no_match
        httpMethod: POST
        contentHandling: CONVERT_TO_TEXT
        type: aws_proxy
        uri: 
          Fn::Sub: arn:aws:apigateway:${AWS::Region}:lambda:path/2015-03-31/functions/arn:aws:lambda:${AWS::Region}:${AWS::AccountId}:function:${ApplicationPrefix}-lambda/invocations

  /api/v1/checkMembership:
    get:
      summary: Get ACM membership status for a given Illinois NetID.
      operationId: checkMembership
      parameters:
        - in: query
          name: netId
          schema:
            type: string
          required: true
          description: Illinois NetID to check.
      responses:
        200:
          description: OK

      x-amazon-apigateway-auth:
        type: NONE

      x-amazon-apigateway-integration:
        responses:
          default:
            statusCode: 200
        passthroughBehavior: when_no_match
        httpMethod: POST
        contentHandling: CONVERT_TO_TEXT
        type: aws_proxy
        uri: 
          Fn::Sub: arn:aws:apigateway:${AWS::Region}:lambda:path/2015-03-31/functions/arn:aws:lambda:${AWS::Region}:${AWS::AccountId}:function:${ApplicationPrefix}-lambda/invocations

  /api/v1/checkExternalMembership:
    get:
      summary: Get external list membership status for a given Illinois NetID.
      operationId: checkExternalMembership
      parameters:
        - in: query
          name: netId
          schema:
            type: string
          required: true
          description: Illinois NetID to check.
        - in: query
          name: list
          schema:
            type: string
          required: true
          description: List to check against
      responses:
        200:
          description: OK

      x-amazon-apigateway-auth:
        type: NONE

      x-amazon-apigateway-integration:
        responses:
          default:
            statusCode: 200
        passthroughBehavior: when_no_match
        httpMethod: POST
        contentHandling: CONVERT_TO_TEXT
        type: aws_proxy
        uri: 
          Fn::Sub: arn:aws:apigateway:${AWS::Region}:lambda:path/2015-03-31/functions/arn:aws:lambda:${AWS::Region}:${AWS::AccountId}:function:${ApplicationPrefix}-lambda/invocations

  
  /api/v1/checkout/session:
    get:
      summary: Get a Stripe checkout session for purchasing an ACM membership.
      operationId: getMembershipCheckoutSession
      parameters:
        - in: query
          name: netid
          schema:
            type: string
          required: true
          description: Illinois NetID to create session for.
      responses:
        200:
          description: OK

      x-amazon-apigateway-auth:
        type: NONE

      x-amazon-apigateway-integration:
        responses:
          default:
            statusCode: 200
        passthroughBehavior: when_no_match
        httpMethod: POST
        contentHandling: CONVERT_TO_TEXT
        type: aws_proxy
        uri: 
          Fn::Sub: arn:aws:apigateway:${AWS::Region}:lambda:path/2015-03-31/functions/arn:aws:lambda:${AWS::Region}:${AWS::AccountId}:function:${ApplicationPrefix}-lambda/invocations
  
  /api/v1/provisioning/member:
    post:
      summary: Provision a member from Stripe webhook
      operationId: postMemberProvision
      responses:
        200:
          description: OK

      x-amazon-apigateway-auth:
        type: NONE

      x-amazon-apigateway-integration:
        responses:
          default:
            statusCode: 200
        passthroughBehavior: when_no_match
        httpMethod: POST
        contentHandling: CONVERT_TO_TEXT
        type: aws_proxy
        uri: 
          Fn::Sub: arn:aws:apigateway:${AWS::Region}:lambda:path/2015-03-31/functions/arn:aws:lambda:${AWS::Region}:${AWS::AccountId}:function:${ApplicationPrefix}-lambda/invocations
  